refactor: svg import 시 불필요한 ?react 쿼리 제거
